+++ Source/EventFlow.Elasticsearch/Extensions/EventFlowOptionsExtensions.cs
@@ -23,12 +23,10 @@
 using System;
 using System.Linq;
 using Elasticsearch.Net;
-using EventFlow.Aggregates;
-using EventFlow.Configuration;
-using EventFlow.Core;
 using EventFlow.Elasticsearch.ReadStores;
 using EventFlow.Extensions;
 using EventFlow.ReadStores;
+using Microsoft.Extensions.DependencyInjection;
 using Nest;
 
 namespace EventFlow.Elasticsearch.Extensions
@@ -70,8 +68,8 @@ namespace EventFlow.Elasticsearch.Extensions
         {
             return eventFlowOptions.RegisterServices(sr =>
                 {
-                    sr.Register(f => elasticClientFactory(), Lifetime.Singleton);
-                    sr.Register<IReadModelDescriptionProvider, ReadModelDescriptionProvider>(Lifetime.Singleton, true);
+                    sr.AddSingleton(_ => elasticClientFactory());
+                    sr.AddSingleton<IReadModelDescriptionProvider, ReadModelDescriptionProvider>();
                 });
         }
 
@@ -94,24 +92,11 @@ namespace EventFlow.Elasticsearch.Extensions
                 .UseReadStoreFor<IElasticsearchReadModelStore<TReadModel>, TReadModel, TReadModelLocator>();
         }
 
-        [Obsolete("Use the simpler method UseElasticsearchReadModel<TReadModel> instead.")]
-        public static IEventFlowOptions UseElasticsearchReadModelFor<TAggregate, TIdentity, TReadModel>(
-            this IEventFlowOptions eventFlowOptions)
-            where TAggregate : IAggregateRoot<TIdentity>
-            where TIdentity : IIdentity
+        private static void RegisterElasticsearchReadStore<TReadModel>(IServiceCollection collection)
             where TReadModel : class, IReadModel
         {
-            return eventFlowOptions
-                .RegisterServices(RegisterElasticsearchReadStore<TReadModel>)
-                .UseReadStoreFor<TAggregate, TIdentity, IElasticsearchReadModelStore<TReadModel>, TReadModel>();
-        }
-
-        private static void RegisterElasticsearchReadStore<TReadModel>(
-            IServiceRegistration serviceRegistration)
-            where TReadModel : class, IReadModel
-        {
-            serviceRegistration.Register<IElasticsearchReadModelStore<TReadModel>, ElasticsearchReadModelStore<TReadModel>>();
-            serviceRegistration.Register<IReadModelStore<TReadModel>>(r => r.Resolver.Resolve<IElasticsearchReadModelStore<TReadModel>>());
+            collection.AddTransient<IElasticsearchReadModelStore<TReadModel>, ElasticsearchReadModelStore<TReadModel>>();
+            collection.AddTransient<IReadModelStore<TReadModel>>(r => r.GetRequiredService<IElasticsearchReadModelStore<TReadModel>>());
         }
     }
 }